refactor: useMovies 훅 추가
