+++ src/App.tsx
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import styled from "styled-components";
-import type { Movie } from "./types/movie";
 import { Header } from "./components/Header";
 import { MovieForm } from "./components/MovieForm";
 import { MovieList } from "./components/MovieList";
 import { useThemeStore } from "./store/themeStore";
+import { useMovies } from "./hooks/useMovies";
 
 const AppContainer = styled.div<{ $theme: any }>`
   background: ${props => props.$theme.background};
@@ -23,58 +23,8 @@ const ErrorMessage = styled.div<{ $theme: any }>`
 `;
 
 function App() {
-  const { themeName, currentTheme, toggleTheme } = useThemeStore();
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
-  useEffect(() => {
-    const fetchMovies = async () => {
-      setIsLoading(true);
-      try {
-        const res = await fetch("/api/movies");
-        if (!res.ok) throw new Error("영화 데이터를 불러오지 못했습니다.");
-        const data: Movie[] = await res.json();
-        setMovies(data);
-      } catch (err: unknown) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError(String(err));
-        }
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    fetchMovies();
-  }, []);
-
-  const handleAddMovie = async (newMovie: Omit<Movie, "id">) => {
-    try {
-      setIsLoading(true);
-      setError(null);
-      
-      const res = await fetch("/api/movies", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newMovie),
-      });
-
-      if (!res.ok) throw new Error("영화를 추가하지 못했습니다.");
-
-      const savedMovie: Movie = await res.json();
-      setMovies((prev) => [...prev, savedMovie]);
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        setError(err.message);
-      } else {
-        setError(String(err));
-      }
-      throw err; // 에러를 다시 throw해서 MovieForm에서 처리할 수 있도록
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const { currentTheme } = useThemeStore();
+  const { movies, error, isLoading, addMovie } = useMovies();
 
   return (
     <AppContainer $theme={currentTheme}>
@@ -85,7 +35,7 @@ function App() {
         </ErrorMessage>
       )}
       <MovieForm 
-        onAddMovie={handleAddMovie}
+        onAddMovie={addMovie}
         isLoading={isLoading}
       />
       <MovieList
